requirements.txt
gphoto2
opencv-python-headless
PyQt6
Pillow
